requirements.txt
torch==1.7.1+cu110
torchvision==0.8.2+cu110
numpy==1.19.4
tqdm==4.54.1
Pillow==8.0.1
future==0.18.2
faiss-gpu==1.7.0
typing==3.7.4.3
functools==0.5
gdown==3.13.0
